dom/base/test/test_bug1187157.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=789315
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 789315</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=789315">Mozilla Bug 789315</a>
<form id="a"><input name="b" type="file"/></form>

<script type="text/javascript">
  var obj = new FormData(document.getElementById('a')).get('b');
  is(obj, "", "We want an empty string.");

</script>

</body>
</html>
